pyproject.toml
[project]
name = "bsky-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "atproto>=0.0.58",
    "fastapi>=0.115.8",
    "flask>=3.1.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
